:ref: extracted code into common modules for easier re-usability
