+++ console/module/data/src/main/java/org/eclipse/kapua/app/console/module/data/servlet/DataExporterServlet.java
@@ -18,11 +18,13 @@ import org.eclipse.kapua.KapuaUnauthenticatedException;
 import org.eclipse.kapua.app.console.module.api.setting.ConsoleSetting;
 import org.eclipse.kapua.app.console.module.api.setting.ConsoleSettingKeys;
 import org.eclipse.kapua.app.console.module.api.shared.util.GwtKapuaCommonsModelConverter;
-import org.eclipse.kapua.app.console.module.data.shared.util.KapuaGwtDataModelConverter;
+import org.eclipse.kapua.app.console.module.api.shared.util.KapuaGwtCommonsModelConverter;
 import org.eclipse.kapua.locator.KapuaLocator;
 import org.eclipse.kapua.service.datastore.MessageStoreFactory;
 import org.eclipse.kapua.service.datastore.MessageStoreService;
 import org.eclipse.kapua.service.datastore.internal.mediator.MessageField;
+import org.eclipse.kapua.service.datastore.internal.setting.DatastoreSettings;
+import org.eclipse.kapua.service.datastore.internal.setting.DatastoreSettingsKey;
 import org.eclipse.kapua.service.datastore.model.DatastoreMessage;
 import org.eclipse.kapua.service.datastore.model.MessageListResult;
 import org.eclipse.kapua.service.datastore.model.query.MessageQuery;
@@ -50,6 +52,8 @@ public class DataExporterServlet extends HttpServlet {
     private static final Logger LOG = LoggerFactory.getLogger(DataExporterServlet.class);
 
     private static final KapuaLocator LOCATOR = KapuaLocator.getInstance();
+
+    private static final DatastoreSettings DATASTORE_SETTINGS = LOCATOR.getComponent(DatastoreSettings.class);
 
     private static final MessageStoreFactory MESSAGE_STORE_FACTORY = LOCATOR.getFactory(MessageStoreFactory.class);
 
@@ -134,7 +138,7 @@ public class DataExporterServlet extends HttpServlet {
             query.setPredicate(predicate);
             MessageListResult result;
             long totalCount = 0;
-            totalCount = KapuaGwtDataModelConverter.countEsDataCapped(MESSAGE_STORE_SERVICE, query);
+            totalCount = KapuaGwtCommonsModelConverter.countEsDataCapped(MESSAGE_STORE_SERVICE, query, DATASTORE_SETTINGS.getInt(DatastoreSettingsKey.MAX_RESULT_WINDOW_VALUE));
 
             long totalOffset = 0;
             query.setLimit(250);
